requirements.txt
Django==2.2.28
mixer==7.1.2
Pillow==9.0.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
django-environ==0.8.1
